[CloseOpportunity] Use helper function to work out if opportunity can be closed
